+++ pkg/kube/handler_test.go
@@ -2457,5 +2457,4 @@ func TestImportKube(t *testing.T) {
 				testCase.expectedCode, rr.Code)
 		}
 	}
-
 }
